+++ cockpit/src/components/output-block.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function OutputBlock({children}) {
+  return <pre className="nas-pre">{children}</pre>;
+}