+++ cockpit/src/index.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import {createRoot} from "react-dom/client";
 import "@patternfly/patternfly/patternfly.css";
 import "@patternfly/patternfly/patternfly-addons.css";
+import "./cockpit-dark-theme.js";
 import App from "./app.jsx";
 import "./app.scss";